Started a trial branch to remove the LoadCase class.
Split Element into its own file and test files to allow for easier
modification of the Element class (easier to read through shorter files)
